+++ src/main/java/dev/handyshulkers/HandyShulkers.java
@@ -1,5 +1,6 @@
 package dev.handyshulkers;
 
+import dev.handyshulkers.config.HandyShulkersConfig;
 import net.fabricmc.api.ModInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,6 +19,7 @@ public class HandyShulkers implements ModInitializer {
 
 	@Override
 	public void onInitialize() {
+		HandyShulkersConfig.load();
 		LOGGER.info("Handy Shulkers loaded! Shulker boxes now behave like bundles.");
 	}
 }